### 2020-01-08 0e255fb
add GPL3 note to README
