### 2020-01-18 #2030 @whedon
[REVIEW]: FGTpartitioner: A rapid method for parsimonious delimitation of ancestry breakpoints in large genome-wide SNP datasets
**Submitting author:** @tkchafin (<a href="http://orcid.org/0000-0001-8687-5905">Tyler Chafin</a>)
**Repository:** <a href="https://github.com/tkchafin/FGTpartitioner" target ="_blank">https://github.com/tkchafin/FGTpartitioner</a>
**Version:** v1.0.1
