### 2020-01-23 #2030 @YichaoOU
The documentation and manuscript are well written. Installation and test cases are easy to follow.

In my case, `./FGTpartitioner.py -h` does not work, but `python FGTpartitioner.py -h` works.

The GitHub repo provides examples but the documentation didn't specifically provide test commands such as, `python FGTpartitioner.py -v examples/example_1.vcf.gz`

Minor comments:

1.  What is the time complexity for your algorithm? Remember in the Wang 2010 paper, they claimed to have O(mn) and O(nm^2) in their proposed DP algorithms. It is OK that you don't write a formal algorithmic complexity analysis, but I think it is better to show the complexity of your algorithm.

2. Did you use any publicly available data to verify your method?

Thanks,
Yichao
