### 2020-01-24 tkchafin/FGTpartitioner#2 @dangeles
FGTpartitioner.py -h fails
https://github.com/openjournals/joss-reviews/issues/2030

I followed the setup instructions on the README. Installation seemingly went flawlessly. When I typed:

`./FGTpartitioner.py -h`

I get the following error when using python 3.7.6:
```
  File "./FGTpartitioner.py", line 73
    print("Diploid resolution strategy: ", end="")
                                              ^
SyntaxError: invalid syntax
```

However, `python FGTpartitioner.py -h` works. I think the header in the script is wrong---it should be `#!/usr/bin/env`, not `#!/usr/bin/python`.
