### 2020-01-27 tkchafin/FGTpartitioner#2 @tkchafin
Fixed in latest commit- although depends on the location of your python install (e.g. still fails for me since my Python is in ~/anaconda3/bin). Code now correctly points to the default system python though, and I added a note in the README
