### 2020-01-27 #2030 @dangeles
This is a neat algorithm that addresses an important problem for non-canonical model organism genomics. The README and paper are well-written, and in general the code is well-documented and nicely written.

Doc-strings are missing from all defined function in `FGTpartitioner.py`. It would improve the code significantly to add short doc-strings explaining the input, output and purpose of each function.

Two example datasets are provided, but to the best of my ability, it appears they give identical results, and no guidance is given to the user as to why two examples are provided. The `example_noHet` is included without a `.tbi` file (probably a good teaching moment). On the other hand, additional files are provided within the examples folder that could probably be ignored or documented somewhere.

Best,
David
