### 2020-02-03 6fe2af1
Delete benchmarking.md
