### 2020-02-03 c03b5aa
Delete remake_example1.sh
