+++ examples/example_README.md
@@ -1,0 +1,6 @@
+This is a description of the files in this directory:
+- example_1_genome.fasta : Example FASTA-formatted "genome" for use with example_1.vcf* files
+- example_1.vcf : Uncompressed .VCF file 
+- example_1.vcf.gz : Compressed .VCF file produced using bgzip (see README.md file for exact command)
+- example_1.vcf.gz.tbi : Tabix index file for the .vcf.gz (see README.md)
+- example_2_alignment.fasta : Example FASTA-formatted "genome alignment" file, with monomorphic bases all coded as "N" to make the variants more visible
