+++ examples/example_README.md
@@ -3,4 +3,4 @@ This is a description of the files in this directory:
 - example_1.vcf : Uncompressed .VCF file 
 - example_1.vcf.gz : Compressed .VCF file produced using bgzip (see README.md file for exact command)
 - example_1.vcf.gz.tbi : Tabix index file for the .vcf.gz (see README.md)
-- example_2_alignment.fasta : Example FASTA-formatted "genome alignment" file, with monomorphic bases all coded as "N" to make the variants more visible
+- example_1_alignment.fasta : Example FASTA-formatted "genome alignment" file, showing the full alignment from which the example_1_genome.fasta and example_1.vcf* files were derived. This file is included because some similar programs (e.g. Rmincutter.pl) take a full alignment as input. 
