### 2020-02-03 7a8931b
added docstrings
