### 2020-02-03 #2030 @tkchafin
@lpantano
- Added information to the README with further benchmarking information including publicly-available data (accession numbers for NCBI SRA) which was used for comparison/ validation against some other published methods.
- Fixed the path issue causing the "-h" issue both reviewers noted
- Added an Examples section to README given explicit commands to verify functionality
- Added a README file to examples/ directory stating what each file is, and standardized the file naming in that directory
- Added docstrings throughout
Most recent commit (https://github.com/tkchafin/FGTpartitioner/commit/aa8fe7c35ed0887cfd7621175a919e907f60053d) is up to date on all of these changes
