### 2020-02-04 #2030 @lpantano
@tkchafin, thank you! Then, last two steps: 1) is the version still v1.0.0, 2) can you create a zenodo DOI link where the title and the author match exactly the paper information and give it to us?
Thanks!
