### 2020-02-04 11fd03e
Update FGTpartitioner.py
