+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+## Changelog
+
+# Releases
+
+v1.0.1
+
+- Added benchmarking to documentation
+- Docstrings throughout code
+- Fixed small bug in Python path
+
+v1.0.0
+
+- Initial release
