### 2020-02-06 #2030 @arfon
> @openjournals/dev I'm not able to figure out what is going wrong with the paper here. Could you help?

Apologies, I accidentally introduced a regression last night. This should be fixed now.
